requirements.txt
langgraph>=0.0.40
langsmith>=0.0.50
langchain-openai>=0.1.0
langchain-core>=0.2.0
composio>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.0.0
discord.py>=2.3.2
